+++ desktop/src/main/java/bisq/desktop/primary/main/content/trade/components_poc/ProtocolSelection.java
@@ -15,7 +15,7 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.desktop.primary.main.content.trade.components;
+package bisq.desktop.primary.main.content.trade.components_poc;
 
 import bisq.account.protocol_type.ProtocolTypeUtil;
 import bisq.account.protocol_type.TradeProtocolType;
