+++ desktop/src/main/java/bisq/desktop/primary/main/content/trade/multisig/old/createOffer/MultiSigCreateOfferController.java
@@ -27,10 +27,10 @@ import bisq.desktop.common.view.InitWithDataController;
 import bisq.desktop.common.view.Navigation;
 import bisq.desktop.common.view.NavigationTarget;
 import bisq.desktop.components.controls.MarketSelection;
-import bisq.desktop.primary.main.content.trade.components.AmountPriceGroup;
-import bisq.desktop.primary.main.content.trade.components.DirectionSelection;
-import bisq.desktop.primary.main.content.trade.components.PaymentSelection;
-import bisq.desktop.primary.main.content.trade.components.ProtocolSelection;
+import bisq.desktop.primary.main.content.trade.components_poc.AmountPriceGroup;
+import bisq.desktop.primary.main.content.trade.components_poc.DirectionSelection;
+import bisq.desktop.primary.main.content.trade.components_poc.PaymentSelection;
+import bisq.desktop.primary.main.content.trade.components_poc.ProtocolSelection;
 import bisq.offer.Direction;
 import bisq.offer.poc.PocOpenOfferService;
 import javafx.collections.SetChangeListener;
