+++ desktop/src/main/java/bisq/desktop/primary/main/content/trade/multisig/old/takeOffer/TakeOfferController.java
@@ -22,8 +22,8 @@ import bisq.common.currency.Market;
 import bisq.desktop.common.view.InitWithDataController;
 import bisq.desktop.common.view.Navigation;
 import bisq.desktop.common.view.NavigationTarget;
-import bisq.desktop.primary.main.content.trade.components.AmountPriceGroup;
-import bisq.desktop.primary.main.content.trade.components.DirectionSelection;
+import bisq.desktop.primary.main.content.trade.components_poc.AmountPriceGroup;
+import bisq.desktop.primary.main.content.trade.components_poc.DirectionSelection;
 import bisq.desktop.primary.main.content.trade.multisig.old.pendingTrades.PendingTradesController;
 import bisq.desktop.primary.main.content.trade.multisig.old.takeOffer.components.TakersPaymentSelection;
 import bisq.offer.Direction;
